requirements.txt
# general  modules
django == 4.1.5
djangorestframework
python-decouple
psycopg2-binary
pillow

# # third party modules
markdown       
django-filter  
drf-yasg[validation]
djangorestframework-simplejwt
djoser
django-cors-headers

# # email third party modules
django-mail-templated

# # background proccess and cache
celery
django_celery_beat
redis
django-redis

# deployment
gunicorn